Add files via upload
Perl script that extracts Cause of Death information and creates triples with annotations.  Two mapping files are needed for this script (COD_mapping_strings.txt and COD_mapping_categories.txt).  Directory information for input files and output triple files is set within the code and must be modified to run on the target platform.